% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_array.R
\name{js_array}
\alias{js_array}
\title{TODO array}
\usage{
js_array(
  .items,
  .min_items = NULL,
  .max_items = NULL,
  .unique_items = NULL,
  ...
)
}
\arguments{
\item{.items}{schema component describing array items}

\item{.min_items}{minimum item count}

\item{.max_items}{maximum item count}

\item{.unique_items}{boolean indicating if items must be unique}

\item{...}{
  Arguments passed on to \code{\link[=js_common_attributes]{js_common_attributes}}
  \describe{
    \item{\code{.element}}{js_schema_component}
    \item{\code{.title}}{title for schema component}
    \item{\code{.description}}{longer description for schema component}
    \item{\code{.comment}}{'$comment' tag for internal comments of schema component}
    \item{\code{.examples}}{TODO doc}
    \item{\code{.enum}}{restrict element by list of valid elements}
    \item{\code{.const}}{restrict element by constant value that must be met}
    \item{\code{.default}}{default value for component if not explicitly set}
    \item{\code{.required}}{logical indicating if element is required}
  }}
}
\description{
TODO array
}
