% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_numeric.R
\name{js_numeric}
\alias{js_numeric}
\title{TODO numeric}
\usage{
js_numeric(
  .multiple_of = NULL,
  .minimum = NULL,
  .maximum = NULL,
  .exclusive_minimum = NULL,
  .exclusive_maximum = NULL,
  ...
)
}
\arguments{
\item{.multiple_of}{force component to be a multiple of that number}

\item{.minimum}{inclusive minimum valid value}

\item{.maximum}{inclusive maximum valid value}

\item{.exclusive_minimum}{exclusive minimum valid value}

\item{.exclusive_maximum}{exclusive maximum valid value}

\item{...}{
  Arguments passed on to \code{\link[=js_common_attributes]{js_common_attributes}}
  \describe{
    \item{\code{.element}}{js_schema_component}
    \item{\code{.title}}{title for schema component}
    \item{\code{.description}}{longer description for schema component}
    \item{\code{.comment}}{'$comment' tag for internal comments of schema component}
    \item{\code{.examples}}{TODO doc}
    \item{\code{.enum}}{restrict element by list of valid elements}
    \item{\code{.const}}{restrict element by constant value that must be met}
    \item{\code{.default}}{default value for component if not explicitly set}
    \item{\code{.required}}{logical indicating if element is required}
  }}
}
\description{
TODO numeric
}
