% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_object.R
\name{js_object}
\alias{js_object}
\title{TODO object}
\usage{
js_object(
  ...,
  .min_properties = NULL,
  .max_properties = NULL,
  .pattern_properties = NULL,
  .additional_properties = NULL
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=js_common_attributes]{js_common_attributes}}
  \describe{
    \item{\code{.element}}{js_schema_component}
    \item{\code{.title}}{title for schema component}
    \item{\code{.description}}{longer description for schema component}
    \item{\code{.comment}}{'$comment' tag for internal comments of schema component}
    \item{\code{.examples}}{TODO doc}
    \item{\code{.enum}}{restrict element by list of valid elements}
    \item{\code{.const}}{restrict element by constant value that must be met}
    \item{\code{.default}}{default value for component if not explicitly set}
    \item{\code{.required}}{logical indicating if element is required}
  }}

\item{.min_properties}{minimum number of properties}

\item{.max_properties}{maximum number of properties}

\item{.pattern_properties}{TODO}

\item{.additional_properties}{boolean indicating if additional properties are allowed}
}
\description{
TODO object
}
