% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_string.R
\name{js_string}
\alias{js_string}
\title{TODO string}
\usage{
js_string(
  .min_length = NULL,
  .max_length = NULL,
  .pattern = NULL,
  .format = c("date-time", "date", "time", "duration", "email", "idn-email", "hostname",
    "idn-hostname", "ipv4", "ipv6", "uuid", "uri", "uri-reference", "iri",
    "iri-reference", "uri-template", "json-pointer", "relative-json-pointer", "regex"),
  ...
)
}
\arguments{
\item{.min_length}{non negative integer describing minimum length of the string}

\item{.max_length}{non negative integer describing maximum length of the string}

\item{.pattern}{scalar string describing regular expression pattern that the string must match}

\item{.format}{character vector of strings describing format of the string}

\item{...}{
  Arguments passed on to \code{\link[=js_common_attributes]{js_common_attributes}}
  \describe{
    \item{\code{.element}}{js_schema_component}
    \item{\code{.title}}{title for schema component}
    \item{\code{.description}}{longer description for schema component}
    \item{\code{.comment}}{'$comment' tag for internal comments of schema component}
    \item{\code{.examples}}{TODO doc}
    \item{\code{.enum}}{restrict element by list of valid elements}
    \item{\code{.const}}{restrict element by constant value that must be met}
    \item{\code{.default}}{default value for component if not explicitly set}
    \item{\code{.required}}{logical indicating if element is required}
  }}
}
\description{
TODO string
}
