% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber.schema-package.R
\docType{package}
\name{plumber.schema-package}
\alias{plumber.schema}
\alias{plumber.schema-package}
\title{plumber.schema: Implementation of JSON Schema for plumber}
\description{
Implementation of JSON Schema for plumber. This package provides a function to generate a plumber router from a JSON Schema. The router can be used to validate requests against the schema and to generate OpenAPI documentation.
}
\author{
\strong{Maintainer}: Nils Schnakenberg \email{schnakenberg.nils@gmail.com}

}
\keyword{internal}
