% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber.schema-package.R
\docType{package}
\name{jsonschema-package}
\alias{jsonschema}
\alias{jsonschema-package}
\title{jsonschema: Implementation of JSON Schema in R}
\description{
Implementation of JSON Schema in R. JSON Schema is a vocabulary that allows you to annotate and validate JSON documents. For more information about JSON Schema itself, please see \url{https://json-schema.org/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jsonschema.nschnakenberg.eu/}
}

}
\author{
\strong{Maintainer}: Nils Schnakenberg \email{schnakenberg.nils@gmail.com}

}
\keyword{internal}
